Added blog posts

+++ shop/services/manage/Blog/CategoryManageService.php
@@ -6,14 +6,17 @@ use shop\entities\Blog\Category;
 use shop\entities\Meta;
 use shop\forms\manage\Blog\CategoryForm;
 use shop\repositories\Blog\CategoryRepository;
+use shop\repositories\Blog\PostRepository;
 
 class CategoryManageService
 {
     private $repository;
+    private $posts;
 
-    public function __construct(CategoryRepository $repository)
+    public function __construct(CategoryRepository $repository, PostRepository $posts)
     {
         $this->repository = $repository;
+        $this->posts = $posts;
     }
 
     public function get($id): ?Category
@@ -60,6 +63,9 @@ class CategoryManageService
     public function remove($id): void
     {
         $category = $this->get($id);
+        if($this->posts->existsByCategory($id)){
+            throw new \DomainException('Cannot delete category with posts.');
+        }
         $this->repository->remove($category);
     }
 }
